Doctors details page initial commit
