+++ frontend/lib/pages/doctor_details.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+
+class DoctorDetailsPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      
+    );
+  }
+}
